Refactor geometry handling, improve viewer reliability, and update dev app
- **Core**:
  - Add strokeDashArray support to AnnotationStyle and fabric-utils
  - Refactor getGeometryFromFabricObject to fabric-utils
- **Tools**:
  - Moved annotation deletion logic to `BaseTool` for centralized handling.
  - Fix race condition in multi-delete
  - Optimize SelectTool to avoid unnecessary ID generation
- **Viewer**:
  - Enhanced `ViewerCell` to automatically detect image types (DZI vs. simple image) for more reliable loading.
  - Switched to `canvas.insertAt` for improved annotation z-ordering during updates.
- **Dev App**:
  - Updated example images and tool configurations in `App.tsx`.
  - Improved `index.html` layout styling for full-screen usage.
- **Cleanup**: Removed unused imports and types across the codebase.
