+++ dev/index.html
@@ -1,12 +1,38 @@
 <!DOCTYPE html>
 <html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Image Annotator — Dev</title>
-  </head>
-  <body>
-    <div id="app"></div>
-    <script type="module" src="./App.tsx"></script>
-  </body>
+
+<head>
+  <meta charset="UTF-8" />
+  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+  <title>Image Annotator — Dev</title>
+  <style>
+    * {
+
+      box-sizing: border-box;
+
+    }
+
+    body {
+      width: 100vw;
+      height: 100vh;
+    }
+
+    html,
+    body {
+      margin: 0;
+      padding: 0;
+    }
+
+    #app {
+      width: 100%;
+      height: 100%;
+    }
+  </style>
+</head>
+
+<body>
+  <div id="app"></div>
+  <script type="module" src="./App.tsx"></script>
+</body>
+
 </html>
