+++ src/core/tools/base-tool.ts
@@ -1,7 +1,8 @@
-import { Rect, Circle, Line, Polyline, Polygon, util, FabricObject } from 'fabric';
+import { FabricObject } from 'fabric';
 import type { FabricOverlay } from '../../overlay/fabric-overlay.js';
-import { Annotation, AnnotationType, Point, Geometry, AnnotationStyle, ImageId, AnnotationContextId, AnnotationId, ToolConstraint, createAnnotationId } from '../types.js';
+import { Annotation, AnnotationType, Point, AnnotationStyle, ImageId, AnnotationContextId, AnnotationId, ToolConstraint, createAnnotationId } from '../types.js';
 import { generateId } from '../../utils/id.js';
+import { getGeometryFromFabricObject } from '../fabric-utils.js';
 
 /** Framework-agnostic callbacks that tools use to interact with application state */
 export interface ToolCallbacks {
@@ -59,6 +60,12 @@ export abstract class BaseTool implements AnnotationTool {
     this.callbacks = callbacks;
   }
 
+  onKeyDown(event: KeyboardEvent): void {
+    if (event.key === 'Delete' || event.key === 'Backspace') {
+        this.deleteSelected();
+    }
+  }
+
   deactivate(): void {
     this.cancel();
     this.overlay = null;
@@ -69,8 +76,24 @@ export abstract class BaseTool implements AnnotationTool {
   abstract onPointerDown(event: PointerEvent, imagePoint: Point): void;
   abstract onPointerMove(event: PointerEvent, imagePoint: Point): void;
   abstract onPointerUp(event: PointerEvent, imagePoint: Point): void;
-  onKeyDown(_event: KeyboardEvent): void {}
   abstract cancel(): void;
+
+  private deleteSelected() {
+    if (!this.callbacks || !this.imageId || !this.overlay) return;
+    const activeObjects = this.overlay.canvas.getActiveObjects().slice();
+    if (!activeObjects || activeObjects.length === 0) return;
+
+    // Discard selection first to prevent Fabric from errors when objects are removed
+    this.overlay.canvas.discardActiveObject();
+    this.overlay.canvas.requestRenderAll();
+
+    for (const obj of activeObjects) {
+        const annotationId = (obj as AnnotatedFabricObject).annotationId;
+        if (annotationId) {
+            this.callbacks.deleteAnnotation(annotationId, this.imageId);
+        }
+    }
+  }
 }
 
 export function createAnnotationFromFabricObject(
@@ -90,72 +113,4 @@ export function createAnnotationFromFabricObject(
     geometry,
     style,
   };
-}
-
-function getGeometryFromFabricObject(obj: FabricObject, type: AnnotationType): Geometry | null {
-  if (type === 'rectangle' && obj instanceof Rect) {
-    const width = obj.width * obj.scaleX;
-    const height = obj.height * obj.scaleY;
-    const left = obj.left;
-    const top = obj.top;
-
-    return {
-      type: 'rectangle',
-      origin: { x: left, y: top },
-      width: width,
-      height: height,
-      rotation: obj.angle,
-    };
-  }
-
-  if (type === 'circle' && obj instanceof Circle) {
-      const radius = obj.radius * Math.max(Math.abs(obj.scaleX), Math.abs(obj.scaleY));
-      const center = obj.getCenterPoint();
-      return {
-          type: 'circle',
-          center: { x: center.x, y: center.y },
-          radius: radius,
-      };
-  }
-
-  if (type === 'line' && obj instanceof Line) {
-      const matrix = obj.calcTransformMatrix();
-      const cx = (obj.x1 + obj.x2) / 2;
-      const cy = (obj.y1 + obj.y2) / 2;
-      const p1 = util.transformPoint({ x: obj.x1 - cx, y: obj.y1 - cy }, matrix);
-      const p2 = util.transformPoint({ x: obj.x2 - cx, y: obj.y2 - cy }, matrix);
-      return {
-          type: 'line',
-          start: { x: p1.x, y: p1.y },
-          end: { x: p2.x, y: p2.y },
-      };
-  }
-
-  if (type === 'point' && obj instanceof Circle) {
-      const center = obj.getCenterPoint();
-      return {
-          type: 'point',
-          position: { x: center.x, y: center.y },
-      };
-  }
-
-  if (type === 'path') {
-      // Polygon extends Polyline in Fabric, so instanceof Polyline matches both
-      if (obj instanceof Polyline) {
-          const matrix = obj.calcTransformMatrix();
-          const pathOffset = obj.pathOffset || { x: 0, y: 0 };
-          const points = (obj.points || []).map(p => {
-              const centeredP = { x: p.x - pathOffset.x, y: p.y - pathOffset.y };
-              const tp = util.transformPoint(centeredP, matrix);
-              return { x: tp.x, y: tp.y };
-          });
-          return {
-              type: 'path',
-              points: points,
-              closed: obj instanceof Polygon,
-          };
-      }
-  }
-
-  return null;
 }
